refactor(theoldllm): extract proxy guards into dedicated module
